% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parasiteclearance.R
\name{clean_profile}
\alias{clean_profile}
\title{Clean one parasite-time profile}
\usage{
clean_profile(profile, detection_limit = 16)
}
\arguments{
\item{profile}{Data frame with columns `time` and `parasitaemia`.}

\item{detection_limit}{Detection limit in parasites per microlitre.}
}
\value{
Data frame with cleaning flags and `used` column.
}
\description{
Clean one parasite-time profile
}
